pyproject.toml
[project]
name = "magisterium-archiver"
version = "0.1.0"
description = "Corpus archiver for Catholic magisterial documents"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.27",
    "selectolax>=0.3.21",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "rich>=13.7",
    "click>=8.1",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"
